refactor: move entrypoint to ./cmd/gf
This refactor is necessary for next commit: a WASM entrypoint that uses the
cobra commands.

This WASM entrypoint must not conflict with the current entrypoint.
Therefore, we move the current entrypoint from `.` to `./cmd/gf`
And we declare a new entrypoint `./cmd/wasm` in the next commit

Both entrypoints use the cobra commands declared in lib ./internal/gf
The `rootCmd` has been made public and becomes `RootCmd` so that it's
available to all entrypoint.

This is also in line with golang conventions.
Package `cmd` was located in ./cmd but was named `cmd` instead of `main`
and did not contain a main() entrypoint